requirements.txt
ultralytics
opencv-python
streamlit
streamlit-folium
folium
numpy
pillow
matplotlib
requests
